requirements.txt
langchain
python-dotenv
langchain-community
pypdf
beautifulsoup4
arxiv
PyMuPDF
langchain-text-splitters
langchain-openai
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
streamlit
